+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/SøknadRouter.tsx
@@ -16,7 +16,8 @@ import MedlemskapStep from './steps/medlemskap/MedlemskapStep';
 import FraværStep from './steps/fravær/FraværStep';
 import OppsummeringStep from './steps/oppsummering/OppsummeringStep';
 import KvitteringPage from '../pages/kvittering/KvitteringPage';
-import FosterbarnStep from './steps/fosterbarn/FosterbarnStep';
+import DineBarnStep from './steps/dine-barn/DineBarnStep';
+import DeltBostedStep from './steps/delt-bosted/DeltBostedStep';
 
 const SøknadRouter = () => {
     const { pathname } = useLocation();
@@ -71,7 +72,8 @@ const SøknadRouter = () => {
         <Routes>
             <Route index element={<VelkommenPage />} />
             <Route path={SøknadStepRoutePath[StepId.VELKOMMEN]} element={<VelkommenPage />} />
-            <Route path={SøknadStepRoutePath[StepId.FOSTERBARN]} element={<FosterbarnStep />} />
+            <Route path={SøknadStepRoutePath[StepId.DINE_BARN]} element={<DineBarnStep />} />
+            <Route path={SøknadStepRoutePath[StepId.DELT_BOSTED]} element={<DeltBostedStep />} />
             <Route path={SøknadStepRoutePath[StepId.SITUASJON]} element={<SituasjonStep />} />
             <Route path={SøknadStepRoutePath[StepId.FRAVÆR]} element={<FraværStep />} />
             <Route path={SøknadStepRoutePath[StepId.LEGEERKLÆRING]} element={<LegeerklæringStep />} />