+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/søknadStepConfig.ts
@@ -1,10 +1,12 @@
 import { SoknadApplicationType, SoknadStepsConfig, soknadStepUtils, StepConfig } from '@navikt/sif-common-soknad-ds';
 import { StepId } from '../types/StepId';
 import { getSøknadStepRoute } from '../utils/søknadRoutesUtils';
+import { Søknadsdata } from '../types/søknadsdata/Søknadsdata';
 
-const getSøknadSteps = (): StepId[] => {
+const getSøknadSteps = (søknadsdata: Søknadsdata): StepId[] => {
     return [
-        StepId.FOSTERBARN,
+        StepId.DINE_BARN,
+        ...(søknadsdata.dineBarn?.harDeltBosted ? [StepId.DELT_BOSTED] : []),
         StepId.SITUASJON,
         StepId.FRAVÆR,
         StepId.LEGEERKLÆRING,
@@ -13,13 +15,13 @@ const getSøknadSteps = (): StepId[] => {
     ];
 };
 
-export const getSøknadStepConfig = (): SoknadStepsConfig<StepId> =>
-    soknadStepUtils.getStepsConfig(getSøknadSteps(), SoknadApplicationType.SOKNAD, (step) => {
+export const getSøknadStepConfig = (søknadsdata: Søknadsdata): SoknadStepsConfig<StepId> =>
+    soknadStepUtils.getStepsConfig(getSøknadSteps(søknadsdata), SoknadApplicationType.SOKNAD, (step) => {
         return getSøknadStepRoute(step);
     });
 
-export const getSøknadStepConfigForStep = (stepId: StepId): StepConfig<StepId> => {
-    const config = getSøknadStepConfig()[stepId];
+export const getSøknadStepConfigForStep = (søknadsdata: Søknadsdata, stepId: StepId): StepConfig<StepId> => {
+    const config = getSøknadStepConfig(søknadsdata)[stepId];
     if (!config) {
         throw `Missing step config ${stepId}`;
     }